requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
authlib==1.4.0
itsdangerous==2.2.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
redis==5.2.1
arq==0.26.1
structlog==24.4.0
python-json-logger==3.2.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
aiofiles==24.1.0
